Support BaiDou, rename `as_` methods, use errors
A `BdtTime` alias has been added for BaiDou Time.

All fallible `as_*` methods have been renamed to `to_*` since `as` is
usually expected to be an infailible narrowing cast in Rust.

All fallible `from_` and `to_` methods now return an error rather than
an option, with the rationale that future `from_` and `to_` methods may
need to return more complex error, which would be inconsistent with the
use of an option in other similar methods (for instance a parsing error
may be returned from a hypothetical `from_str` error).